Add initial bench mark, start refactor of some test elements
