+++ kanidmd/src/lib/idm/mod.rs
@@ -6,7 +6,7 @@ pub(crate) mod event;
 pub(crate) mod group;
 pub(crate) mod mfareg;
 pub(crate) mod radius;
-pub(crate) mod server;
+pub mod server;
 pub(crate) mod unix;
 // mod identity;
