+++ kanidmd/src/lib/lib.rs
@@ -22,9 +22,9 @@ extern crate lazy_static;
 
 // This has to be before 'be' so the import order works
 #[macro_use]
-mod macros;
+pub mod macros;
 mod crypto;
-mod utils;
+pub mod utils;
 #[macro_use]
 mod async_log;
 #[macro_use]
@@ -32,21 +32,21 @@ pub mod audit;
 pub mod be;
 pub mod constants;
 pub mod credential;
-mod entry;
-mod event;
+pub mod entry;
+pub mod event;
 mod filter;
 mod interval;
 pub(crate) mod ldap;
 mod modify;
-mod value;
+pub mod value;
 #[macro_use]
 mod plugins;
 mod access;
 mod actors;
-mod idm;
+pub mod idm;
 mod repl;
 mod schema;
-mod server;
+pub mod server;
 mod status;
 
 pub mod config;
